LayoutTests/http/tests/dom/new-window-can-target-opener-expected.txt

CONSOLE MESSAGE: PASS: New window should have an opener
CONSOLE MESSAGE: PASS: New window should be able to look up opener by name
CONSOLE MESSAGE: PASS: New window should have URL 'http://127.0.0.1:8000/dom/new-window-can-target-opener.html'
Make sure that windows opened via window.open can target their opener

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
